pyproject.toml
[project]
name = "lob_data_engine"
version = "0.2.0"
description = "High-performance LOB data collection and processing engine"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "websockets>=12.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "requests>=2.31.0",
    "typing_extensions>=4.0.0",
    "websocket-client>=1.6.0",
    "eth-account>=0.8.0",
    "msgpack>=1.0.0",
    "eth-utils>=2.0.0",
    "fastparquet>=2023.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["lob_data_engine*"]
exclude = ["tests*", "scripts*", "utils*"]
